+++ frontend/src/app/routes/pages.ts
@@ -1,0 +1,8 @@
+export { ArticleDetailPage } from './pages/ArticleDetailPage';
+export { EditorPage } from './pages/EditorPage';
+export { HomePage } from './pages/HomePage';
+export { LoginPage } from './pages/LoginPage';
+export { NotFoundPage } from './pages/NotFoundPage';
+export { ProfilePage } from './pages/ProfilePage';
+export { RegisterPage } from './pages/RegisterPage';
+export { SettingsPage } from './pages/SettingsPage';
