+++ frontend/src/app/routes/pages/RegisterPage.tsx
@@ -1,0 +1,24 @@
+import { type ReactElement } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useAuth } from '@/app/providers/useAuth';
+import { RegisterForm } from '@/features/auth';
+import { AuthPage } from './AuthPage';
+
+export function RegisterPage(): ReactElement {
+  const { register } = useAuth();
+  const navigate = useNavigate();
+
+  function handleSuccess(): void {
+    navigate('/', { replace: true });
+  }
+
+  return (
+    <AuthPage
+      alternateHref="/login"
+      alternateText="Have an account?"
+      heading="Sign up"
+    >
+      <RegisterForm onSubmit={register} onSuccess={handleSuccess} />
+    </AuthPage>
+  );
+}
